# services/core/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PrevioPLS API",
    "description": "API REST da plataforma Ford Predict & Care (PrevioPLS). Backend SOA para retenção pós-venda. Atende às histórias US01–US07.",
    "contact": {
      "name": "Equipe PrevioPLS — Challenge Ford FIAP 2026"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    },
    {
      "url": "https://api.previopls.example.com",
      "description": "Produção"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "meta",
      "description": "Saúde e descoberta do serviço"
    },
    {
      "name": "clientes",
      "description": "Cadastro D0 da compra — dispara classificação"
    },
    {
      "name": "leads",
      "description": "Leads priorizados consumidos pelo app mobile"
    },
    {
      "name": "auth",
      "description": "Autenticação (JWT)"
    }
  ],
  "paths": {
    "/v1/clientes": {
      "post": {
        "tags": [
          "clientes"
        ],
        "summary": "Cadastra cliente D0 da compra e dispara classificação preditiva",
        "operationId": "cadastrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteCreatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Autentica usuário e retorna JWT",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/leads/{leadId}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Visão 360 — cliente + veículo + script comercial",
        "operationId": "obter",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leads"
        ],
        "summary": "Atualiza status do lead (agendado/recusado/sem-contato)",
        "operationId": "atualizar",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Versão da API e build atual",
        "operationId": "version",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Lista leads filtrados por prioridade/status (paginado)",
        "operationId": "listar",
        "parameters": [
          {
            "name": "prioridade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "critica",
                "alta",
                "media",
                "baixa"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "aberto",
                "agendado",
                "recusado",
                "sem-contato"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeadListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Liveness/readiness — checa conectividade com o banco",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "ClienteCreateRequest": {
        "required": [
          "cpf",
          "nome",
          "regiao",
          "veiculo"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "maxLength": 180,
            "minLength": 2,
            "type": "string"
          },
          "cpf": {
            "pattern": "^\\d{11}$",
            "type": "string"
          },
          "email": {
            "maxLength": 180,
            "minLength": 0,
            "type": "string"
          },
          "telefone": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "regiao": {
            "maxLength": 40,
            "minLength": 2,
            "type": "string"
          },
          "veiculo": {
            "$ref": "#/components/schemas/VeiculoRequest"
          }
        }
      },
      "VeiculoRequest": {
        "required": [
          "ano",
          "concessionariaId",
          "dataCompra",
          "modelo",
          "valorCompra",
          "versao",
          "vin"
        ],
        "type": "object",
        "properties": {
          "modelo": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "versao": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "ano": {
            "maximum": 2100,
            "minimum": 2000,
            "type": "integer",
            "format": "int32"
          },
          "vin": {
            "maxLength": 17,
            "minLength": 17,
            "pattern": "^[A-HJ-NPR-Z0-9]{17}$",
            "type": "string"
          },
          "dataCompra": {
            "type": "string",
            "format": "date"
          },
          "valorCompra": {
            "minimum": 0.0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "concessionariaId": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ClienteCreatedResponse": {
        "type": "object",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/ClienteResponse"
          },
          "leadId": {
            "type": "string",
            "format": "uuid"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "fiel",
              "abandono",
              "esquecido",
              "economico"
            ]
          },
          "scoreRisco": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ClienteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "regiao": {
            "type": "string"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "fiel",
              "abandono",
              "esquecido",
              "economico"
            ]
          },
          "scoreRisco": {
            "type": "number",
            "format": "double"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "classificadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "veiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VeiculoResponse"
            }
          }
        }
      },
      "VeiculoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelo": {
            "type": "string"
          },
          "versao": {
            "type": "string"
          },
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "vin": {
            "type": "string"
          },
          "dataCompra": {
            "type": "string",
            "format": "date"
          },
          "valorCompra": {
            "type": "number"
          },
          "concessionariaId": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "senha"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 180,
            "minLength": 0,
            "type": "string"
          },
          "senha": {
            "maxLength": 128,
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "LeadPatchRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "aberto",
              "agendado",
              "recusado",
              "sem-contato"
            ]
          },
          "observacao": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "LeadDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scoreRisco": {
            "type": "number",
            "format": "double"
          },
          "prioridade": {
            "type": "string",
            "enum": [
              "critica",
              "alta",
              "media",
              "baixa"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "aberto",
              "agendado",
              "recusado",
              "sem-contato"
            ]
          },
          "scriptOferta": {
            "type": "string"
          },
          "observacao": {
            "type": "string"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "cliente": {
            "$ref": "#/components/schemas/ClienteResponse"
          },
          "veiculo": {
            "$ref": "#/components/schemas/VeiculoResponse"
          }
        }
      },
      "VersionResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "build": {
            "type": "string"
          }
        }
      },
      "LeadListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "veiculoId": {
            "type": "string",
            "format": "uuid"
          },
          "nomeCliente": {
            "type": "string"
          },
          "modeloVeiculo": {
            "type": "string"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "fiel",
              "abandono",
              "esquecido",
              "economico"
            ]
          },
          "scoreRisco": {
            "type": "number",
            "format": "double"
          },
          "prioridade": {
            "type": "string",
            "enum": [
              "critica",
              "alta",
              "media",
              "baixa"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "aberto",
              "agendado",
              "recusado",
              "sem-contato"
            ]
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LeadListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeadListItemResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "components": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}